firs

+++ app/src/main/java/com/example/kolodeznapp/API/JokeApi.kt
@@ -17,7 +17,7 @@ interface JokeApi {
 
     fun addJoke(joke: Joke)
 
-    fun getJokeById(id: Int): Joke
+    fun getJokeById(id: Long): Joke
 
     fun getAllJoke(): Array<Joke>
 }
